+++ src/mgen/backends/llvm/runtime/mgen_vec_double.h
@@ -1,0 +1,1 @@
+../../c/runtime/mgen_vec_double.h